Split futures_unordered.rs
